Realized while trying to isolation-test our "view helpers" that the isolation feedback wasn't helpful.
In response to that pain, I could tell that rather
than "view helpers" we really had scar tissue to
distance ourselves from the details of bootstrap's
popover. So here I renamed it to make that purpose
more clear.

You can see in the previous commit that the test
is *not* an isolation test b/c it calls through to 
the actual library. This is appropriate for wrapper
code, because it verifies any assumptions we make
about the behavior of that library, and mocking offers
little value since it can't provide us feedback, 
since we can't change popover's API without forking
bootstrap.js