Adds initial specs for the rendering of the sign-up view.
Note that I changed the production code slightly.

This is an interesting point. I wrapped the view's
template option with a function. I didn't *have*
to do this, but I like to go out of my way to avoid
direct manipulation of the subject-under-test in 
any way that isn't typical for public use.

Instead, by wrapping it in a function, Backbone will
defer template discovery until render time, and I'm
able to edit the depended-on template function instead.

It's a subtle difference but has proven to be a
useful habit over the years.