adds a test for createAccount.
on the @model this introduces:

I always start with the bare minimum test double I
can and then add onto it as the subject needs. If
the dependency (@model) is just an empty javascript
object with a method on it, future maintainers can 
easily see that the dependency is lightweight.

However, often a view will depend on more of the
guts in Backbone.Model (particularly attrs and events)
and in that case I may opt for using a vanilla
Backbone.Event object or a Backbone.Model. To insist
on a completely fake test double can obscure the 
test with a lot of setup code that really just 
reimplements aspects of backbone in order to stub.